Update SolCx and SolKz benchamrks in tests folder
